[Refactor] apply SRP, separate the metadata-related APIs into a brand new class `FileTreeMetadataController`
